Add new settings sections for General, GIF, Mouse Clicks, Video, and Screenshot
- Implement GeneralSettingsSection for output and advanced settings.
- Create GifSettingsSection for GIF capture settings and options.
- Introduce MouseClicksSettingsSection to manage mouse click visuals.
- Add VideoSettingsSection for video recording settings and options.
- Implement ScreenshotSettingsSection for screenshot capture settings.
- Refactor SettingsView to utilize new section views for better organization.
- Add ProSettingsSection for managing Pro subscription details.
- Create MouseClickOverlayControls for customizing mouse click visuals.
- Add helper functions and bindings in SettingsViewHelpers.
- Ensure compatibility with App Store version by using conditional compilation.
