+++ TinyClips/Views/Settings/GeneralSettingsSection.swift
@@ -1,0 +1,101 @@
+import SwiftUI
+
+struct GeneralSettingsSection: View {
+    @ObservedObject var settings: CaptureSettings
+    @ObservedObject var launchAtLogin: LaunchAtLoginManager
+    let sparkleController: SparkleController
+    let openWindow: OpenWindowAction
+    let showDisableDockWarning: Binding<Bool>
+    let chooseSaveDirectory: () -> Void
+    let resetAllSettings: () -> Void
+    let showInDockBinding: Binding<Bool>
+
+    var body: some View {
+        Section("Output") {
+#if APPSTORE
+            VStack(alignment: .leading, spacing: 6) {
+                Text("Default locations: Screenshots/GIFs → Pictures/TinyClips, Videos → Movies/TinyClips")
+                    .font(.caption)
+                    .foregroundStyle(.secondary)
+
+                HStack {
+                    Text(settings.saveDirectoryDisplayPath.isEmpty ? "Using default folders" : settings.saveDirectoryDisplayPath)
+                        .lineLimit(1)
+                        .truncationMode(.middle)
+
+                    Spacer()
+
+                    Button("Browse…") {
+                        chooseSaveDirectory()
+                    }
+
+                    if settings.hasCustomSaveDirectory {
+                        Button("Reset") {
+                            // Reset handled in parent
+                        }
+                    }
+                }
+            }
+#else
+            HStack {
+                TextField("Save to", text: $settings.saveDirectory)
+                    .textFieldStyle(.roundedBorder)
+                Button("Browse…") {
+                    chooseSaveDirectory()
+                }
+            }
+#endif
+            VStack(alignment: .leading, spacing: 6) {
+                TextField("File name template", text: $settings.fileNameTemplate)
+                    .textFieldStyle(.roundedBorder)
+
+                HStack(spacing: 10) {
+                    Button("Classic") {
+                        settings.fileNameTemplate = "TinyClips {date} at {time}"
+                    }
+                    .buttonStyle(.link)
+
+                    Button("Type + Date") {
+                        settings.fileNameTemplate = "{type} {date} at {time}"
+                    }
+                    .buttonStyle(.link)
+
+                    Button("Date First") {
+                        settings.fileNameTemplate = "{date} {time} {type}"
+                    }
+                    .buttonStyle(.link)
+                }
+                .font(.caption)
+
+                Text("Tokens: {app}, {type}, {date}, {time}, {datetime}")
+                    .font(.caption)
+                    .foregroundStyle(.secondary)
+
+                Text("Preview: \(SaveService.shared.namingPreview(for: .screenshot))")
+                    .font(.caption2)
+                    .foregroundStyle(.secondary)
+                    .lineLimit(1)
+                    .truncationMode(.middle)
+            }
+
+            Toggle("Show in Finder after save", isOn: $settings.showInFinder)
+            Toggle("Show notification after save", isOn: $settings.showSaveNotifications)
+        }
+
+        Section("Advanced") {
+            Toggle("Launch at login", isOn: Binding(
+                get: { launchAtLogin.isEnabled },
+                set: { launchAtLogin.setEnabled($0) }
+            ))
+            Toggle("Show TinyClips in Dock (enables ⌘⇥)", isOn: showInDockBinding)
+                .help("When enabled, TinyClips appears in Command-Tab and can participate in normal app/window switching.")
+            Toggle("Always capture main display", isOn: $settings.alwaysCaptureMainDisplay)
+                .help("Skip the display picker when multiple monitors are connected")
+            Toggle("Include TinyClips in captures", isOn: $settings.includeTinyClipsInCapture)
+                .help("For developer/demo use. When enabled, TinyClips windows can appear in screenshots, recordings, and window selection.")
+            Button("Reset All Settings to Defaults…") {
+                resetAllSettings()
+            }
+        }
+    }
+}
