+++ TinyClips/Views/Settings/SettingsViewHelpers.swift
@@ -1,0 +1,5 @@
+import Foundation
+import SwiftUI
+import AppKit
+
+// Place shared helpers, bindings, and utility functions here for SettingsView and its subviews.
